+++ src/langchain_google_cloud_sql_pg/engine.py
@@ -22,8 +22,9 @@ from typing import TYPE_CHECKING, Awaitable, Dict, List, Optional, TypeVar
 import aiohttp
 import google.auth  # type: ignore
 import google.auth.transport.requests  # type: ignore
-from google.cloud.sql.connector import Connector, create_async_connector
-from sqlalchemy import text  # Column,
+from google.cloud.sql.connector import Connector
+from sqlalchemy import MetaData, Table, text
+from sqlalchemy.exc import InvalidRequestError
 from sqlalchemy.ext.asyncio import AsyncEngine, create_async_engine
 
 from .version import __version__
@@ -86,7 +87,7 @@ class Column:
             raise ValueError("Column data_type must be type string")
 
 
-class PostgreSQLEngine:
+class PostgresEngine:
     """A class for managing connections to a Cloud SQL for Postgres database."""
 
     _connector: Optional[Connector] = None
@@ -110,7 +111,7 @@ class PostgreSQLEngine:
         database: str,
         user: Optional[str] = None,
         password: Optional[str] = None,
-    ) -> PostgreSQLEngine:
+    ) -> PostgresEngine:
         # Running a loop in a background thread allows us to support
         # async methods from non-async environments
         loop = asyncio.new_event_loop()
@@ -139,7 +140,7 @@ class PostgreSQLEngine:
         password: Optional[str] = None,
         loop: Optional[asyncio.AbstractEventLoop] = None,
         thread: Optional[Thread] = None,
-    ) -> PostgreSQLEngine:
+    ) -> PostgresEngine:
         if bool(user) ^ bool(password):
             raise ValueError(
                 "Only one of 'user' or 'password' were specified. Either "
@@ -190,7 +191,7 @@ class PostgreSQLEngine:
         database: str,
         user: Optional[str] = None,
         password: Optional[str] = None,
-    ) -> PostgreSQLEngine:
+    ) -> PostgresEngine:
         return await cls._create(
             project_id,
             region,
@@ -201,7 +202,7 @@ class PostgreSQLEngine:
         )
 
     @classmethod
-    def from_engine(cls, engine: AsyncEngine) -> PostgreSQLEngine:
+    def from_engine(cls, engine: AsyncEngine) -> PostgresEngine:
         return cls(engine, None, None)
 
     async def _aexecute(self, query: str, params: Optional[dict] = None):
@@ -226,12 +227,12 @@ class PostgreSQLEngine:
         return result_fetch
 
     def _execute(self, query: str, params: Optional[dict] = None):
-        return self.run_as_sync(self._aexecute(query, params))
+        return self._run_as_sync(self._aexecute(query, params))
 
     def _fetch(self, query: str, params: Optional[dict] = None):
-        return self.run_as_sync(self._afetch(query, params))
+        return self._run_as_sync(self._afetch(query, params))
 
-    def run_as_sync(self, coro: Awaitable[T]) -> T:
+    def _run_as_sync(self, coro: Awaitable[T]) -> T:
         if not self._loop:
             raise Exception("Engine was initialized async.")
         return asyncio.run_coroutine_threadsafe(coro, self._loop).result()
@@ -279,7 +280,7 @@ class PostgreSQLEngine:
         overwrite_existing: bool = False,
         store_metadata: bool = True,
     ) -> None:
-        return self.run_as_sync(
+        return self._run_as_sync(
             self.ainit_vectorstore_table(
                 table_name,
                 vector_size,
@@ -303,7 +304,7 @@ class PostgreSQLEngine:
         await self._aexecute(create_table_query)
 
     def init_chat_history_table(self, table_name) -> None:
-        return self.run_as_sync(
+        return self._run_as_sync(
             self.ainit_chat_history_table(
                 table_name,
             )
@@ -350,7 +351,7 @@ class PostgreSQLEngine:
         metadata_json_column: str = "langchain_metadata",
         store_metadata: bool = True,
     ) -> None:
-        return self.run_as_sync(
+        return self._run_as_sync(
             self.ainit_document_table(
                 table_name,
                 content_column,
@@ -359,3 +360,34 @@ class PostgreSQLEngine:
                 store_metadata,
             )
         )
+
+    async def _aload_table_schema(
+        self,
+        table_name: str,
+    ) -> Table:
+        """
+        Load table schema from existing table in PgSQL database.
+        Returns:
+            (sqlalchemy.Table): The loaded table.
+        """
+        metadata = MetaData()
+        async with self._engine.connect() as conn:
+            try:
+                await conn.run_sync(metadata.reflect, only=[table_name])
+            except InvalidRequestError as e:
+                raise ValueError(f"Table, {table_name}, does not exist: " + str(e))
+
+        table = Table(table_name, metadata)
+        # Extract the schema information
+        schema = []
+        for column in table.columns:
+            schema.append(
+                {
+                    "name": column.name,
+                    "type": column.type.python_type,
+                    "max_length": getattr(column.type, "length", None),
+                    "nullable": not column.nullable,
+                }
+            )
+
+        return metadata.tables[table_name]