+++ CHANGELOG.md
@@ -1,1 +1,19 @@
 # Changelog
+
+## 0.1.0 (2024-02-28)
+
+
+### Features
+
+* Add CloudSQL Postgresql chatmessagehistory with Integration Tests ([#23](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/issues/23)) ([3ab9d4e](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/commit/3ab9d4eeeb7fd99c4693ee697fb31a2ad9343872))
+* Add indexing methods ([#21](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/issues/21)) ([8eae440](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/commit/8eae4406e41f234ef3c6a24621926c3f5c4555cb))
+* Add PostgreSQL loader ([#49](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/issues/49)) ([ada45ec](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/commit/ada45ec3089254966e444d11c5c22f73b881d03b))
+* Add PostgreSQLEngine ([#13](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/issues/13)) ([b181f65](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/commit/b181f658c2e769c74aefc6a53f587ca4a75682db))
+* Add Vector Store ([#14](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/issues/14)) ([f3e1127](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/commit/f3e11276a69bf239d852e494eede37ed86b1b361))
+
+
+### Documentation
+
+* Add chatmessagehistory docs ([#48](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/issues/48)) ([5f5df1d](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/commit/5f5df1d0790dd0a90110a1c765a4f445c083267a))
+* Add the code lab notebook ([#36](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/issues/36)) ([ab7cbe4](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/commit/ab7cbe4d0554a2a80a32e7feb7b4fc5c773ee379))
+* Add vectorstore docs ([#22](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/issues/22)) ([6c41df2](https://github.com/googleapis/langchain-google-cloud-sql-pg-python/commit/6c41df2f51c7b185d8d1b53ad6b12e42f32de224))